% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree_manual.R
\name{cutree_manual}
\alias{cutree_manual}
\title{Cut a hierarchical tree.}
\usage{
cutree_manual(tree, hc = NULL, hf = NULL)
}
\arguments{
\item{tree}{(return value of hclust()), hc (double, under which height to cluster trees), hf (double, above which height to ignore).}
}
\value{
a list (integer + a list of vertex).
}
\description{
This function groups close enough stocks into vertices (<= hc). It assign stocks between hc to hf to single-point-vertex (hc< h <= hf). It groups all stocks that are not closely correlated to each other into the last vertex (> hf).
This function returns a list, with first element being num_vertices (integer), second element points_in_vertex (a vector).
}
