% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_params.R
\name{find_params}
\alias{find_params}
\title{Automatically produce hc and hf}
\usage{
find_params(tree)
}
\arguments{
\item{tree}{(return value of hclust())}
}
\value{
a list of hc and hf
}
\description{
This function automatically generate parameters to cut the hierachical tree. 
hc: the height under which there are 1/3 points. create a histogram, find the mid point of the bin that left bins of this bin contains 1/3 of the points.
hf: the mid point of the bin that is the first empty bin on the right of the hc bin in the histogram after.
}
