% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_to_net.R
\name{igraph_to_net}
\alias{igraph_to_net}
\title{Generate net information.
This function takes all information of an igraph object and combine them.}
\usage{
igraph_to_net(g, name = NULL, group = NULL, size = NULL,
  comp = NULL, what = "both")
}
\arguments{
\item{"g"}{(an igraph object), "name" of vertices (vector), "group" (a vector of integers range from 1 to N), "size" of vertices (vector), "comp"_onents of the piechart (a list of vectors).}
}
\value{
a data frame of either nodes or links or a list of both data frames.
}
\description{
Generate net information.
This function takes all information of an igraph object and combine them.
}
