% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_to_json.R
\name{net_to_json}
\alias{net_to_json}
\title{Translate net information (R data frame) to JSON file.
This function takes all information of an igraph object and combine them.}
\usage{
net_to_json(net, file_loc = NULL)
}
\arguments{
\item{"net"}{(a data frame containing nodes and links information).  "file_loc" which is the output file with full path.}
}
\value{
json format data.
}
\description{
Translate net information (R data frame) to JSON file.
This function takes all information of an igraph object and combine them.
}
